Allow populating on nested arrays of objects
Allow populating on nested array of objects like key[0].name, key[1].name

If keyThere includes a dot, we're looking for a nested prop. This checks if that nested prop is an array.

If it's an array, we assume it to be an array of objects.

It splits the key only on the first dot which allows populating on nested keys inside the array of objects.